added tensor optimizer, legs function, numpy/torch backend support and some other stuff such as xtest for testing

+++ torch_hamiltonians.py
@@ -1,12 +1,12 @@
 import numpy as np
 import torch
 
-szi = torch.cuda.DoubleTensor([[1., 0], [0, -1.]])
-sxi = torch.cuda.DoubleTensor([[0, 1.], [1., 0]])
-syi = torch.cuda.DoubleTensor([[0, -1.],[1., 0]])
-id = torch.cuda.DoubleTensor([[1., 0],[0, 1.]])
-s_plusi = torch.cuda.DoubleTensor([[0,1],[0,0]])
-s_minusi = torch.cuda.DoubleTensor([[0,0],[1,0]])
+szi = torch.cuda.FloatTensor([[1., 0], [0, -1.]])
+sxi = torch.cuda.FloatTensor([[0, 1.], [1., 0]])
+syi = torch.cuda.FloatTensor([[0, -1.],[1., 0]])
+id = torch.cuda.FloatTensor([[1., 0],[0, 1.]])
+s_plusi = torch.cuda.FloatTensor([[0,1],[0,0]])
+s_minusi = torch.cuda.FloatTensor([[0,0],[1,0]])
 
 
 def ising_transverse_x(x,l=1.,j=1.0,):
